Better handling of runtime panics
Instead of abusing Rust's panicking mechanism the VM now uses Result
types where applicable. This makes it much easier to gracefully
terminate the VM while displaying a runtime stack trace, without having
to mess with Rust panic handlers. Panics caused by the Inko runtime
(e.g. the use of an incorrect instruction operand) are called "runtime
panics".

When a runtime panic is triggered the VM will print a stack trace and
terminate. The stack trace contains a variety of colours (that can be
disabled by setting CLICOLOR=0) to make it easier to read the
information. For now the colours of the stack trace are hard-coded, but
they may be configurable in the future.

Fixes #87