Fixed looking up locals in parent bindings
Creating an additional binding when running SetBlock would result in the
VM not always being able to look up the right local variable in the
right binding. Removing this additional binding (and making a few small
changes to the compiler) solves this issue while also reducing the
number of allocated Binding structures.