Use a custom Arc for strings
Rust's own Arc supports weak references, and checking for those when
dropping an Arc adds enough overhead that this becomes visible when
finalising a lot of objects. To work around this we introduce
ArcWithoutWeak: a simplified version of Arc that doesn't support weak
references. Using this structure we can improve finalisation performance
by 1.3x compared to using Rust's Arc for strings.