Rework GC scheduling and process cleanups
This commit does the following:

1. The logic used for determining when to collect garbage has been
   drastically changed.

2. Processes that have finished executing are now cleaned up in a
   separate thread.

These two changes are discussed below.

== Scheduling Changes

The old approach to scheduling garbage collection was pretty simple:
when X or more blocks have been allocated we schedule a collection. At
the end of the collection we increase the threshold so we don't keep
scheduling at the same interval.

While simple to implement this approach would result in fewer
collections over time, resulting in larger heaps (due to the thresholds
getting incremented on every collection). This could result in a program
using much more memory than necessary.

In the new setup we still schedule a collection after a certain number
of blocks has been allocated, but the logic used for incrementing this
threshold is pretty different. Instead of always incrementing this
threshold we only do so when _after_ reclaiming all blocks we still use
a certain percentage of the threshold (90% by default). This ensures we
only increase the heap size when we (may) actually need the space,
instead of always doing this.

Per this commit increasing the mature generation's size will also
increase the young generation's size. The idea is that increasing the
young generation in this case will hopefully lead to fewer objects being
promoted prematurely due to fewer garbage collection cycles being
scheduled.

== Process Cleanups

Previously when a process would finished it would clean up its memory in
the same thread that was executing it (or whatever other thread removed
the last reference to the "RcProcess" structure). Given enough memory
this could lead to an application thread being blocked for hundreds of
milliseconds.

To work around this a process is scheduled for a final "finished"
collection. This collection happens in one of the GC threads and will
simply reclaim and finalize all memory the process was using. This won't
reduce the total time it takes to perform this work, but it does ensure
tihs won't affect application threads.