Cram more meta data in block headers
Since we reserve the first line (= 128 bytes) of a block we can use this
data for more than just a pointer to the block. In this commit we move
the bucket pointer, hole count, and fragmentation status to the block
header. This won't save that much space (up to 24 bytes per block), but
it does allow us to potentially add more fields in the future without
further increasing the memory usage.