Schedule entire Immix blocks for finalization
Instead of scheduling individual pointers for finalization we now
schedule entire Immix blocks. By scheduling entire blocks we remove the
need for gathering and storing individual pointers, reducing the amount
of memory necessary for finalization.

To reduce GC pause timings we don't wait for blocks to be finalized
before resuming a process. Not waiting for the blocks to be finalized
can easily reduce pause timings by a factor of 10. To make this possible
the allocator will first finalize a block that is supposed to be
finalized before allocating into it. This is done using a combination of
atomic operations and locking to reduce the overhead this adds as much
as possible.

In this setup both the mutator and the various finalizer threads will
compete for blocks to finalize. In most cases the finalizer threads will
have done most of the work by the time a process is resumed, but
sometimes a mutator will also have to finalize a block (e.g. if many
blocks have been scheduled).

This fixes https://gitlab.com/yorickpeterse/inko/issues/86