Set up basic STDIN/STDOUT/STDERR runtime methods
These methods are provided in three different modules:

* std::stdio::stdin
* std::stdio::stdout
* std::stdio::stderr

These modules in turn implement the Read/Write traits from the std::io
module. To make writing to STDOUT/STDERR easier the write/print methods
don't throw any errors. The most common error is a broken pipe error and
ignoring those should be perfectly fine as this shouldn't affect the
behaviour of the program. We may revise this in the future if this is
considered too harmful.