Added time formatting and parsing methods
This is where the "time" crate is starting to show it's ugly side. For
example, parsing the string "2018" using the format "%Y" will result in
the day of the year being set to 0, instead of 1. This is because the
"time" crate (and the underlying C functions) start the day of the year
with 0, whereas ISO 8601 starts it at 1.

A workaround would be to fix the Tm structure returned before wrapping
it in a DateTime, but another alternative might be to just use the
"chrono" crate after all. I'll need to give this some thought to see
what solution we end up going with.