Removed Array capacity/reserve instructions
These are redundant because reserving space can be done by just
inserting a value at an index beyond the length. For example, to reserve
four values you'd insert a Nil at index 3. This does result in Nil
values being used to fill up empty slots, but this doesn't require any
more space so it's not a big deal.