Support converting optional types to non optionals
This adds support for converting optional types (e.g. ?T) to their
non-optional equivalents (T in this case). This is useful for cases
where you can guarantee the value is never Nil. For example,
std::os.platform uses this because the VM never returns a platform index
greater than the number of known platform names.

Converting optional types does not generate any runtime checks and as
such should be used with extreme care. In the future I hope to extend
the compiler so it can do this automatically when using "if_true" or "if
true:".