Fixed tracing of the remembered set
When preparing for a young collection we would reset the line maps of
all mature blocks. The idea behind this was to ensure that any mature
objects in the remembered set would be unmarked and thus traced again.
Unfortunately this code _only_ reset the used lines map, completely
ignoring the marks map. As a result after the first collection the
remembered mature objects were left as marked, preventing them from
being traced in following collection cycles. This in turn could lead to
young objects not being marked, resulting in them being overwritten.

To solve all this we now reset both the line and mark maps of all
remembered objects. We do this by simply iterating over the remembered
set instead of iterating over all mature blocks, reducing the time
necessary.