Added support for method type requirements
An example use case would be Array.==: this method should only be
available if all values in the Array implement the Equal trait.
Requiring the Equal trait to be implemented up front (along with any
other traits some methods may require) means potentially having to
implement a lot of unused traits _just_ so you can store a value in the
array.

Using method type requirements we can set method specific requirements,
ensuring the method is only available when all requirements are method.
For Array.== this means we'd define it as follows:

    impl Equal for Array!(T) {
      def ==(other: Self) -> Boolean where T: Equal {
        ...
      }
    }

The code necessary to make this work is a total hack, like most of the
compiler. Chances are it's buggy and might be changed a lot in the
future, but for now it should be good enough.