Added std::marker
This module will be used to make the compiler a bit simpler by checking
for the implementation of certain marker traits, instead of defining
custom objects/methods/etc. For example, instead of using a "Nil" type
to make it compatible with Optional we can just check if
std::marker::Optional is implemented.