Copy when receiving a message
Prior to this commit whenever we'd receive a message, we would simply
take it as-is. This could result in the mailbox collector collecting an
object, even when still referenced from the young or mature generation.

To work around this, we now copy the object upon receiving. To reduce
the memory footprint this has, we move over data where possible. For
example, when receiving an Array we move the native Rust Vec, instead of
creating a new Vec from scratch. To further reduce the memory footprint,
we now also garbage collector the mailbox much more frequently.