Added process.channel and docs
The method `process.channel` can be used to spawn and communicate with a
process, in a way similar to using channels. A benefit of this approach
is that the compiler can ensure the right types of messages are sent and
received, removing the need for dynamic typing.

To illustrate this, let's say you want to spawn a process that waits for
a "ping" message, then sends "pong" back to process 0. Using
`process.spawn` you could do this as follows:

    import std::process
    import std::reflection

    let pid = process.spawn {
      let message = process.receive

      reflection.instance_of?(message, String).if_true {
        (message as String) == 'ping'
          .if_true {
            process.send(0, 'pong')
          }
      }
    }

    process.send(pid, 'ping')

    process.receive # => 'pong'

If we are comfortable the spawned process only receives instances of
String, then we can reduce this down to the following:

    import std::process

    let pid = process.spawn {
      let message = process.receive as String

      message == 'ping'
        .if_true {
          process.send(0, 'pong')
        }
    }

    process.send(pid, 'ping')

    process.receive # => 'pong'

Using `process.channel` we can write the following:

    import std::process

    let sender = process.channel!(String) lambda (receiver) {
      let message = receiver.receive

      message == 'ping'
        .if_true {
          process.send(0, 'pong')
        }
    }

    sender.send('ping')

    process.receive # => 'pong'

While this requires roughly the same amount of code, we don't need to
manually cast any types. We also get to enjoy type safety, as in the
above example the compiler would not allow code such as
`sender.send(10)`.

Fixes https://gitlab.com/yorickpeterse/inko/issues/99