Formatting of objects for debugging purposes
When developing or debugging a program, we may want to display the
contents of an object in a human-readable way. In Ruby this can be done
by sending "inspect" to an object. This commit adds a similar, albeit a
bit more flexible implementation.

The module std::format defines some types and methods that can be used
for building an object formatter. The module std::inspect in turn
uses these types to implement "inspect" for various core types. This
allows us to print a human-readable representation of Inko objects. Some
examples:

    10.inspect           # => '10'
    'hello'.inspect      # => '"hello"'
    [10, 20, 30].inspect # => '[10, 20, 30]'

    object Person {
      def init(name: String, age: Integer) {
        let @name = name
        let @age = age
      }
    }

    let person = Person.new(name: 'Alice', age: 28)

    person.inspect # => Person { @name = "Alice", age = 28 }

"inspect" is implemented for all objects through "Object.inspect", but
objects are free to redefine this method where necessary.

With this change comes the removal of Array.to_string. Converting an
Array to a regular String doesn't make much sense, since it's possible
its values can't be reasonably serialised to text. It's also confusing
to have "to_string" and "inspect" do the same thing, when both methods
serve different purposes:

* "inspect" is meant for debugging purposes.
* "to_string" is meant to produce output the program may want to store
  or use again.