Fix index slice generation in the VM
This changes the way array/string slices are generated by the VM. For
strings we now operate on the number of characters, instead of bytes.
Negative indexes are now also handled better, and should no longer
overflow if either the index or the list of values in a slice is too
great.

Fixes #109