Turn Boolean back into a regular object
Defining Boolean as a Trait was a nice idea, but it was too limiting.
For example, the Inspect trait is defined after Boolean, meaning you
couldn't pass True or False to an argument expecting an Inspect, since
Boolean did not define Inspect as a required trait.

To work around this we turn Boolean back into a regular object, and make
True and False instances of this object. This allows us to refine these
objects whenever necessary, at the cost of having to implement said code
for Boolean, True, False.