Fix Float equality in the VM
The float-cmp crate provides an ApproxEqUlps trait. This trait seems to
treat two NAN values as equal, even though the IEEE-754 specification
states that no two NAN values are equal. To work around this, we first
check if we're dealing with a NAN on the left-hand or right-hand side,
then we perform the actual comparison.