Improve the arch detection script
This now spits out something similar to Rust's target value, e.g.
x86_64-linux-gnu. We're not using "gcc -dumpmachine" because it
appears to produce quite different values on different Linux
distributions. For example:

* Arch linux: x86_64-pc-linux-gnu
* Alpine: x86_64-alpine-linux-musl
* Debian: x86_64-linux-gnu

We exclude the vendor, producing an actual triplet, since it's not
terribly useful (at least so far).