Support building on stable Rust
With Rust 1.28 stabilising the std::alloc crate, we can finally allow
building of IVM on stable Rust. This does require disabling some nightly
only features that improve performance, but for development environments
this should not pose any issues.

To handle this automatically, nightly features are disabled by default.
These nightly-only features are enabled automatically when a nightly
version of Cargo is used. This allows users to easily compile IVM, while
still being able to (easily) opt-in to the various nightly features.