Always set "self" in a lambda
Previously we would treat it as an argument, only setting it when left
out. When we send "call" to a lambda, the first argument passed is the
lambda itself. This would result in "self" being set to the lambda,
because our override would only kick in if the argument was left out.

To solve this, we _always_ set self to the module object in a lambda,
even if an explicit argument is provided.