Rework std::test to use panics instead of throws
This removes the need for using `try assert.something`, instead allowing
developers to write just `assert.something`. This in turn makes it
easier and more natural to write unit tests.

From a philosophical standpoint, using panics is also more correct. An
assertion failure indicates a bug in the code, and using try/else is not
meant to be used for such cases.