Use linked lists for process mailboxes
A LinkedList is quite a bit smaller than a VecDeque, and in theory
should use less memory as it doesn't need to maintain a buffer. Since
all we ever do is pushing and popping, performance should be roughly the
same. Moving messages from the external to internal list may also be a
bit faster, as appending one LinkedList to another is an O(1) operation.