Fix parsing of arguments without parenthesis
This would break whenever passing an Array. For example, this would
fail:

    foo ['bar'], 10

Because "bracket_open" ("[") was not marked as a valid start token for a
value, the parser would throw an error. Adding this token to the list of
allowed tokens wasn't trivial though, as this would then confuse the
parser when parsing code such as this:

    foo[10]

This would be parsed as:

    foo([10])

Instead of:

    foo.[](10)

To solve this, when parenthesis are not provided, the first argument
_must_ be preceded by a space. This ensures we can distinguish between
`foo[10]` and `foo [10]`.