Parse trailing blocks as arguments
When sending a message using parentheses, if the expression is followed
by a block (on the same line as the closing parenthesis), this block is
treated as the final argument. This means that this:

    [10, 20, 30].each() do (number) {
      ...
    }

Is parsed as this:

    [10, 20, 30].each(do (number) {
      ...
    })

This is especially useful when passing regular arguments first. For
example, instead of this:

    test.group 'This is the description of the group', {
      ...
    }

You can now write this instead:

    test.group('This is the description of the group') {
      ...
    }

The benefit of this is that it is more consistent when it comes to the
use of parentheses. Prior to this commit, the recommended approach was
to leave out parentheses if the last argument was a block. This however
is a bit inconsistent. For example:

    foo(10, 20, 30)

    foo 10, 20, 30, do (thing) {
      ...
    }

Supporting the trailing block syntax means there is now one case less
where you leave out the parentheses.