Added std::test::assert.true and false
These methods can be used to assert that something is truthy or falsy.
This means that instead of writing this:

    assert.equal(10 == 10, True)

You can now write:

    assert.true(10 == 10)

This commit also makes a change to the compiler to solve a type
parameter error that could occur in certain cases.

Fixes https://gitlab.com/inko-lang/inko/issues/144