Support for type checking nested required traits
This adds compiler support for handling nested required traits. For
example:

    trait A {}
    trait B: A {}
    trait C: B {}

Previously, when passing a C to an A, the compiler would error as it
would only look for directly required traits in C. With this change this
is no longer the case, as we recursively walk through all required
traits to see if A (in the above example) is eventually required.