Refactor std::range so it compiles again
This required various fixes to the compiler, as well as some refactoring
to get rid of the use of Self types in a few places. In particular,
ToRange now takes a type parameter. This makes more sense than using
Self, as technically there is nothing wrong with a range from 0 to 10.0
(Integer..Float).

Having said all that, I consider the current implementation of
std::range::Range to not be ideal. I will most likely try to rewrite it
much more drastically, hopefully getting rid of the Successor and
Compare requirements, or at least relaxing some of the type requirements
a bit.