Fixed passing of excessive arguments
When passing a single argument to a rest argument, the rest argument
would be set directly to the given value, not an array of it. This
commit fixes the logic for passing excessive arguments to ensure the
rest argument is always an array, regardless of the number of passed
arguments.