Pinning of processes to threads
This commits adds the ability to pin a process to a particular OS thread
in a thread pool. This is useful for the FFI, as certain C functions or
data structures require to be run on a specific thread. For example,
libc's "errno" variable uses thread-local storage. This means that if we
want to run a function that uses it and read "errno", we _have_ to
ensure both operations are performed on the same thread.

To support this, each Pool structure now has a Worker structure, which
stores the thread ID of that worker. This ID can then be stored in a
process, allowing the scheduler to determine which worker should run the
process.

== Process pinning

Pinned processes can not be moved across pools, so the MoveToPool
instruction becomes a noop for a pinned process. We can not panic in
this case, as this would prevent pinned processes from using methods
that try to move a process to the secondary pool.