Simplify setting the default thread counts
This changes the default configuration so that all thread pools default
to using a number of threads equal to the number of logical CPU cores,
instead of using different values for different pools. IVM threads tend
to consume rather little memory. For garbage collection this can also
result in better performance, compared to using a default of 2 threads.