The Platform instruction now returns a string
The "Platform" VM instruction now returns a string, instead of an
integer. This string contains the name of the underlying platform, such
as "windows", "linux", or "unix" for a generic Unix (like) system. This
simplifies the runtime, while being able to detect more platforms. This
in turn can be useful when binding to C using the upcoming FFI API.