Stop using statically linked musl for tests
When musl is statically linked, it doesn't support the use of dlopen().
This would prevent the use of the `std::ffi` module for musl builds.
Instead we'll use the default linkage, which is dynamic. This does mean
that we can no longer provide statically linked musl builds for
development environments, but this is better than not being able to
provide support for `std::ffi`.