Move interpreter code out of the main loop
A lot of code that used to reside directly in the interpreter loop now
resides in separate modules and functions. This makes it a bit easier to
maintain the code, as well as making it easier for a future JIT of sorts
to reuse this code.

Not all instructions have had their logic moved to separate functions.
Typically, if an instruction modifies the instruction pointer its logic
remains in the interpreter loop.