Remove uninitialised params from return types
This changes the compiler so that type parameters that map to
uninitialised type parameters have their instances removed. For type
`Thing!(A -> B)` this means that the mapping of A to B is removed if B
is an uninitialised type parameter.

This change ensures that the compiler can better handle HashMap literals
and other types that may not initialise all type parameters. This in
turn ensures that code such as the following no longer errors:

    let x: HashMap!(String, String) = %[]

Previously this would error because `%[]` was inferred has
`HashMap!(Hash + Equal, V)`, and that type can not safely substitute
`Hashmap!(String, String)`. With this change, `%[]` is basically
inferred as `HashMap!(?, ?)`, allowing the compiler to then infer it
into `HashMap!(String, String)`.