Improve remapping of initialised type parameters
This improves the compiler's support for remapping type parameters to
the appropriate type parameter instance. For example, consider type T
that has type parameter A initialised to type parameter B, and B is
initialised to C. Previously, when looking up the instance of A the
compiler would produce B. As of this commit, the compiler would produce
type C instead. This removes the need for type annotations in various
cases, most notably when working with iterators.

Fixes https://gitlab.com/inko-lang/inko/issues/117